+++ dutchiepay/src/app/_components/_user/_input/AddressInput.jsx
@@ -5,7 +5,7 @@ import '@/styles/user.css';
 
 import { useEffect, useState } from 'react';
 
-import getLocation from '@/app/_components/_user/GetLocation';
+import getLocation from '@/app/_util/GetLocation';
 
 export default function AddressInput({ address, setAddress }) {
   useEffect(() => {
